+++ R/eval.R
@@ -1,14 +1,14 @@
 #' Evaluate a formula
 #'
-#' \code{f_eval_rhs} evaluates the RHS of a formula and \code{f_eval_lhs}
-#' evaluates the LHS. \code{f_eval} is a shortcut for \code{f_eval_rhs} since
+#' \code{tidy_eval_rhs} evaluates the RHS of a formula and \code{tidy_eval_lhs}
+#' evaluates the LHS. \code{tidy_eval} is a shortcut for \code{tidy_eval_rhs} since
 #' that is what you most commonly need.
 #'
 #' If \code{data} is specified, variables will be looked for first in this
 #' object, and if not found in the environment of the formula.
 #'
 #' @section Pronouns:
-#' When used with \code{data}, \code{f_eval} provides two pronouns to make it
+#' When used with \code{data}, \code{tidy_eval} provides two pronouns to make it
 #' possible to be explicit about where you want values to come from:
 #' \code{.env} and \code{.data}. These are thin wrappers around \code{.data}
 #' and \code{.env} that throw errors if you try to access non-existent values.
@@ -18,11 +18,11 @@
 #'   back into the formula. See \code{\link{tidy_quote}()} for more details.
 #' @param data A list (or data frame). \code{data_source} is a generic used to
 #'   find the data associated with a given object. If you want to make
-#'   \code{f_eval} work for your own objects, you can define a method for this
+#'   \code{tidy_eval} work for your own objects, you can define a method for this
 #'   generic.
 #' @export
 #' @examples
-#' f_eval(~ 1 + 2 + 3)
+#' tidy_eval(~ 1 + 2 + 3)
 #'
 #' # formulas automatically capture their enclosing environment
 #' foo <- function(x) {
@@ -31,37 +31,37 @@
 #' }
 #' f <- foo(1)
 #' f
-#' f_eval(f)
+#' tidy_eval(f)
 #'
-#' # If you supply data, f_eval will look their first:
-#' f_eval(~ cyl, mtcars)
+#' # If you supply data, tidy_eval will look their first:
+#' tidy_eval(~ cyl, mtcars)
 #'
 #' # To avoid ambiguity, you can use .env and .data pronouns to be
 #' # explicit:
 #' cyl <- 10
-#' f_eval(~ .data$cyl, mtcars)
-#' f_eval(~ .env$cyl, mtcars)
+#' tidy_eval(~ .data$cyl, mtcars)
+#' tidy_eval(~ .env$cyl, mtcars)
 #'
 #' # Imagine you are computing the mean of a variable:
-#' f_eval(~ mean(cyl), mtcars)
+#' tidy_eval(~ mean(cyl), mtcars)
 #' # How can you change the variable that's being computed?
 #' # The easiest way is "unquote" with !!
 #' # See ?tidy_quote for more details
 #' var <- ~ cyl
-#' f_eval(tidy_quote(mean( !!var )), mtcars)
-#' @name f_eval
-f_eval_rhs <- function(f, data = NULL) {
-  f_eval(f, data)
+#' tidy_eval(tidy_quote(mean( !!var )), mtcars)
+#' @name tidy_eval
+tidy_eval_rhs <- function(f, data = NULL) {
+  tidy_eval(f, data)
 }
-#' @rdname f_eval
+#' @rdname tidy_eval
 #' @export
-f_eval_lhs <- function(f, data = NULL) {
+tidy_eval_lhs <- function(f, data = NULL) {
   f <- f_new(f_lhs(f), env = f_env(f))
-  f_eval(f, data)
+  tidy_eval(f, data)
 }
-#' @rdname f_eval
+#' @rdname tidy_eval
 #' @export
-f_eval <- function(f, data = NULL) {
+tidy_eval <- function(f, data = NULL) {
   if (is_formula(f)) {
     expr <- f_rhs(f)
     env <- f_env(f)